+++ includes/mb/class.coretemp.inc.php
@@ -41,7 +41,7 @@ class Coretemp extends Hwmon
             for ($i = 0; $i < $smp; $i++) {
                 $temp = 0;
                 if (CommonFunctions::executeProgram('sysctl', '-n dev.cpu.'.$i.'.temperature', $temp)) {
-                    $temp = preg_replace('/C/', '', $temp);
+                    $temp = preg_replace('/,/', '.', preg_replace('/C/', '', $temp));
                     $dev = new SensorDevice();
                     $dev->setName("CPU ".($i + 1));
                     $dev->setValue($temp);