+++ includes/output/class.WebpageXSLT.inc.php
@@ -49,6 +49,7 @@ class WebpageXSLT extends WebpageXML implements PSI_Interface_Output
         $domxsl->load($xslfile);
         $xsltproc = new XSLTProcessor;
         $xsltproc->importStyleSheet($domxsl);
+        header("Cache-Control: no-cache, must-revalidate\n");
         echo $xsltproc->transformToXML($domxml);
     }
 }