+++ includes/plugin/class.PSI_Plugin.inc.php
@@ -89,7 +89,7 @@ abstract class PSI_Plugin implements PSI_Interface_Plugin
         if ((!defined('PSI_PLUGIN_'.strtoupper($this->_plugin_name).'_ACCESS')) &&
              (!defined('PSI_PLUGIN_'.strtoupper($this->_plugin_name).'_FILE')) &&
              (!defined('PSI_PLUGIN_'.strtoupper($this->_plugin_name).'_SHOW_SERIAL'))) {
-                $this->global_error->addError("config.ini", "Config for plugin ".$this->_plugin_name." not exist!");
+                $this->global_error->addError("phpsysinfo.ini", "Config for plugin ".$this->_plugin_name." not exist!");
         }
     }