+++ js/jQuery/jquery.treeTable.js
@@ -1,6 +1,6 @@
 /* 
 Copyright: Paul Hanlon
-version 2009-06-22+statefix+spanfix+altfix
+version 2009-06-22+statefix+spanfix+altfix+undefinedfix+ie6cachefix+multilinefix
 Released under the MIT/BSD licence which means you can do anything you want 
 with it, as long as you keep this copyright notice on the page 
 */
@@ -20,17 +20,18 @@ with it, as long as you keep this copyright notice on the page
         if (mapa[ro]){//It's a parent as well. Build it's string and move on to it's children
           pre=(y==yl-1)? opts.blankImg: opts.vertLineImg;
           img=(y==yl-1)? opts.lastOpenImg: opts.openImg;
-          mapb[ro-1] = preStr + '<span class="treeimg"><img src="'+img+'" alt="" class="parimg" id="'+tid+ro+'"></span>';
-          pref = preStr + '<span class="treeimg"><img src="'+pre+'" alt="" class="preimg"></span>';
+          mapb[ro-1] = preStr + '<span class="treeimg" style="background-image: url(\''+pre+'\')"><div style="background-image: url(\''+img+'\')" class="parimg" id="'+tid+ro+'"'+(img=(y==yl-1)? ' last=1': '')+'></div></span>';
+          pref = preStr + '<span class="treeimg preimg" style="background-image: url(\''+pre+'\')"></span>';
           arguments.callee(ro, pref);
         }else{//it's a child
+          pre=(y==yl-1)? opts.blankImg: opts.vertLineImg;
           img = (y==yl-1)? opts.lastLeafImg: opts.leafImg;//It's the last child, It's child will have a blank field behind it
-          mapb[ro-1] = preStr + '<span class="treeimg"><img src="'+img+'" alt="" class="ttimage" id="'+tid+ro+'"></span>';
+          mapb[ro-1] = preStr + '<span class="treeimg" style="background-image: url(\''+pre+'\')"><div style="background-image: url(\''+img+'\')" class="ttimage" id="'+tid+ro+'"></div></span>';
         }
       }
     },
     expandKids = function(num, last){//Expands immediate children, and their uncollapsed children
-      jq("#"+tid+num).attr("src", (last)? opts.lastOpenImg: opts.openImg);//
+      jq("#"+tid+num).css('background-image', "url('"+((last)? opts.lastOpenImg: opts.openImg)+"')");
       for (var x=0, xl=mapa[num].length;x<xl;x++){
         var mnx = mapa[num][x];
         jq("#"+tid+mnx).parents("tr").removeClass("collapsed");
@@ -41,7 +42,7 @@ with it, as long as you keep this copyright notice on the page
     },
     collapseKids = function(num, last){//Recursively collapses all children and their children and change icon
     if (mapa[num]){
-        jq("#"+tid+num).attr("src", (last)? opts.lastShutImg: opts.shutImg);
+        jq("#"+tid+num).css('background-image', "url('"+((last)? opts.lastShutImg: opts.shutImg)+"')");
         for (var x=0, xl=mapa[num].length;x<xl;x++){
           var mnx = mapa[num][x];
           jq("#"+tid+mnx).parents("tr").addClass("collapsed");
@@ -74,13 +75,10 @@ with it, as long as you keep this copyright notice on the page
           return r;
 	};
 
-	var imgtmp = new Image(); //path normalize for ie6
-	imgtmp.src = opts.lastOpenImg;
-	opts.lastOpenImg = imgtmp.src;
-	
-	imgtmp.src = opts.lastShutImg;
-	opts.lastShutImg = imgtmp.src;
-	
+    try { //ie6 flickering fix
+      document.execCommand("BackgroundImageCache", false, true);
+    } catch(err) {}
+
 	for (var x=0,xl=map.length; x<xl;x++){//From map of parents, get map of kids
       num = map[x];
       if (!mapa[num]){
@@ -103,7 +101,7 @@ with it, as long as you keep this copyright notice on the page
     }
     if (collarr.length){
         for (var y=0,yl=collarr.length;y<yl;y++){
-            collapseKids(collarr[y], $("#"+tid+collarr[y]+ ".parimg").attr("src")==opts.lastOpenImg);
+            collapseKids(collarr[y], $("#"+tid+collarr[y]+ ".parimg").attr("last")==1);
         }
     }
     stripe();
@@ -112,11 +110,11 @@ with it, as long as you keep this copyright notice on the page
       jqt.click(function(){
         var num = parseInt(jqt.attr("id").substr(tid.length));//Number of the row
         if (jqt.parents("tr").next().is(".collapsed")){//If the table row directly below is collapsed
-          expandKids(num, (jqt.attr("src")==opts.lastShutImg));//Then expand all children not in collarr
-					if(opts.state){creset(num,true);}//If state is set, store in cookie
+          expandKids(num, jqt.attr("last")==1);//Then expand all children not in collarr
+          if(opts.state){creset(num,true);}//If state is set, store in cookie
         }else{//Collapse all and set image to opts.shutImg or opts.lastShutImg on parents
-          collapseKids(num, (jqt.attr("src")==opts.lastOpenImg));
-					if(opts.state){creset(num,false);}//If state is set, store in cookie
+          collapseKids(num, jqt.attr("last")==1);
+          if(opts.state){creset(num,false);}//If state is set, store in cookie
         }
         stripe();//Restripe the rows
       });